static: switch default port to 11088 in development mode, switch for 80 in Dockerfile
this is an ugly one - i'm still trying to avoid requiring a pre-processing
script for conf files because that gets in the way of having a "live"
development mode, and nginx's support for reading settings from
variables is atrocious. so substitution of real values it is - and 11088
is one character less prone to false positives. and we can't use the higher
port number for the Dockerfile because the nginx docker image has already
EXPOSEd 80, and child images *can't* un-EXPOSE these.